memory: k3-ddr: Add a DDR Controller Driver for K3 family of SoCs
LPDDR4 memories have a temperature sensor feature which can be used by
the processor to update the refresh rate appropriately. The information
from the temperature sensor is stored in a memory register, MR4. The DDR
controller can be setup to automatically read the MR4 register
periodically.

So, add a driver to setup the DDR controller to automatically read the
MR4 register periodically. Also, setup an interrupt handler in the
driver to handle interrupts raised due to the temperature changes.

Signed-off-by: Santhosh Kumar K <s-k6@ti.com>